scripts/requirements.txt
pyarrow==19.0.0
duckdb==1.3.0
sqlglot==26.28.1
pandas==2.3.0
google-cloud-storage==2.18.2
